# Improve performance of uint256