Before
|       Method |     Mean |    Error |   StdDev |
|------------- |---------:|---------:|---------:|
|         Read | 27.78 ns | 0.422 ns | 0.352 ns |
| WriteToArray | 89.02 ns | 1.452 ns | 1.358 ns |
|  WriteToSpan | 20.86 ns | 0.404 ns | 0.338 ns |


After
|       Method |      Mean |     Error |    StdDev |
|------------- |----------:|----------:|----------:|
|         Read |  9.020 ns | 0.1607 ns | 0.1424 ns |
| WriteToArray | 14.387 ns | 0.1589 ns | 0.1487 ns |
|  WriteToSpan | 14.172 ns | 0.1521 ns | 0.1348 ns |

Note that I only do on endian machine, as the logic is too tricky and untestable for bigendian machine. @lontivero @nopara73 you wanted that for a while I think.

@dangershony  I think somebody was also bumping his head on this at one point in stratis?