@lontivero 2020-05-18
Wow. This is really good. Don't forget to share the benchmark project too. It is hard to believe this can be improve further but we also make a lot of comparisons too.