@NicolasDorier 2020-05-18
Sorry, forgot to add the bench file, just did. Test failing on travis, but this is unrelated to this.