@NicolasDorier 2020-05-18
@lontivero to be honest, I am surprised it is so slow. As a comparison, parsing a bech32 takes 5 ns, and base58 10ns... I can't believe those are not doing more stuff...