@NicolasDorier 2020-05-18
I also think we can make things faster by moving from uint to ulong. (probably twice faster)
But I let it to the motivated person :D