@dangershony 2020-05-18
Hi @NicolasDorier  this is great.
At the moment for blockcore I used uint256 improvements made by @mithrilman in his [mithrilshards](https://github.com/MithrilMan/MithrilShards/tree/master/MithrilShards.P2P.Benchmark/Benchmarks/UInt256) project.

See [benchmarks](https://github.com/block-core/blockcore/issues/87) here.

The idea is using 4 ulongs instead of 8 uints