@NicolasDorier 2020-05-20
I think you can win a bit more by using Span cast like I did.